refactor: NotificationMessage 컴포넌트 훅 분리
