+++ frontend/src/components/NotificationMessage/index.tsx
@@ -1,8 +1,5 @@
-import { queryKeys } from 'apis/constants';
-import { useDeleteNotification } from 'apis/pushNotificationApi';
+import { useNotificationMessage } from './useNotificationMessage';
 import { GetNotificationsResponse } from 'apis/pushNotificationApi/type';
-import { useQueryClient } from 'react-query';
-import { useNavigate } from 'react-router-dom';
 import styled, { css } from 'styled-components';
 import { parseTime } from 'utils';
 
@@ -15,34 +12,8 @@ export const NotificationMessage = ({
 }: {
   notifications?: GetNotificationsResponse;
 }) => {
-  const navigate = useNavigate();
   const themeContext = useThemeContext();
-  const queryClient = useQueryClient();
-  const { mutate: deleteNotification } = useDeleteNotification({
-    onSuccess: () => {
-      queryClient.invalidateQueries(queryKeys.getNotifications);
-    },
-  });
-  const handleClickNotification = ({
-    pushNotificationId,
-    pathId,
-    type,
-  }: {
-    pushNotificationId: number;
-    pathId?: number;
-    type?: string;
-  }) => {
-    deleteNotification({ pushNotificationId });
-    if (type === 'comment') {
-      navigate(`/feed/detail/${pathId}`);
-    }
-    if (type === 'challenge') {
-      navigate(`/cycle/detail/${pathId}`);
-    }
-    // TODO : 네비게이트 시킬 라우터 패스, 백엔드로부터 받기
-    // 인증 마감 임박 알릴 경우 -> cycleId 받기 '/cycle/detail/:cycleId',
-    // 댓글 달릴 경우 -> cycleDetailId 받기 '/feed/detail/:cycleDetailId',
-  };
+  const { handleClickNotification } = useNotificationMessage();
 
   return (
     <div>
