+++ frontend/src/components/NotificationMessage/type.ts
@@ -1,0 +1,5 @@
+export interface NotificationHandlerProps {
+  pushNotificationId: number;
+  pathId?: number;
+  type?: string;
+}
